terck: move to utils
