+++ base/src/main/java/org/aya/tyck/order/AyaSccTycker.java
@@ -17,9 +17,9 @@ import org.aya.core.term.Callable;
 import org.aya.core.term.Term;
 import org.aya.generic.util.InterruptException;
 import org.aya.resolve.ResolveInfo;
-import org.aya.terck.CallGraph;
+import org.aya.util.terck.CallGraph;
 import org.aya.terck.CallResolver;
-import org.aya.terck.error.BadRecursion;
+import org.aya.terck.BadRecursion;
 import org.aya.tyck.ExprTycker;
 import org.aya.tyck.StmtTycker;
 import org.aya.tyck.error.CounterexampleError;
